Js update

+++ kotlin-compiler-plugin-swift/src/test/kotlin/com/lightningkite/khrysalis/swift/ExecuteFileTester.ext.kt
@@ -3,62 +3,51 @@ package com.lightningkite.khrysalis.swift
 import com.lightningkite.khrysalis.kotlin.ExecuteFileTester
 import com.lightningkite.khrysalis.kotlin.Libraries
 import com.lightningkite.khrysalis.generic.KotlinTranspileCLP
+import com.lightningkite.khrysalis.util.readInto
+import com.tschuchort.compiletesting.KotlinCompilation
+import com.tschuchort.compiletesting.PluginOption
+import com.tschuchort.compiletesting.SourceFile
 import org.jetbrains.kotlin.util.capitalizeDecapitalize.toLowerCaseAsciiOnly
 import java.io.ByteArrayOutputStream
 import java.io.File
 import java.net.URL
 
-fun swiftInstallation(): File? {
+private var ready = false
+private fun makeReady() {
+    if(ready) return
 
-    val raw = System.getProperty("os.name").toLowerCaseAsciiOnly()
-    return when {
-        raw.contains("win") -> null  // TODO
-        raw.contains("mac") || raw.contains("linux") -> {
-            //Use local installation first
-            val tempFile = File.createTempFile("swiftloc", ".txt")
-            ProcessBuilder().command("which", "swift")
-                .redirectOutput(tempFile)
-                .start()
-                .waitFor()
-            tempFile.readText().trim().takeUnless { it.isEmpty() }?.let { File(it) }
+    //Copy libraries
+    val src = File("../ios-runtime/KhrysalisRuntime/Classes")
+    val dest = swiftTestDir.resolve("Sources/KhrysalisRuntime")
+    if(!dest.exists()) {
+        dest.mkdirs()
+        src.listFiles()!!.forEach {
+            it.copyRecursively(dest.resolve(it.relativeTo(src)), overwrite = true)
         }
-        else -> null
+        dest.resolve("core/UIColor.ext.swift").delete()
+        dest.resolve("android/CGRect+bounds.swift").delete()
     }
+
+    ready = true
 }
 
 val swiftTestDir = File("./testOut")
 fun ExecuteFileTester.swift(sourceFile: File, clean: Boolean): String = caching(sourceFile, clean) {
-
-    //Check swift is installed
-    val swiftExe = swiftInstallation() ?: throw IllegalStateException("Swift not installed")
-
-    //Copy libraries
-    run {
-        val src = File("../ios-runtime/KhrysalisRuntime/Classes")
-        val dest = swiftTestDir.resolve("Sources/KhrysalisRuntime")
-        if(!dest.exists()) {
-            dest.mkdirs()
-            src.listFiles()!!.forEach {
-                it.copyRecursively(dest.resolve(it.relativeTo(src)), overwrite = true)
-            }
-            dest.resolve("core/UIColor.ext.swift").delete()
-            dest.resolve("android/CGRect+bounds.swift").delete()
-        }
-    }
-
+    makeReady()
     val mainFile = swiftTestDir.resolve("Sources/testOut/main.swift")
     mainFile.writeText("print(\"BEGIN PROGRAM\")\nmain()")
     val outputFile = swiftTestDir.resolve("build").resolve(sourceFile.nameWithoutExtension + ".out")
     outputFile.parentFile.mkdirs()
 
-    ProcessBuilder(swiftExe.absolutePath, "run")
+    var output: String = ""
+    ProcessBuilder("swift", "run")
         .directory(swiftTestDir)
         .redirectErrorStream(true)
-        .redirectOutput(outputFile)
         .start()
+        .readInto { output = it }
         .waitFor()
 
-    outputFile.readText().substringAfter("BEGIN PROGRAM").trim()
+    return output.substringAfter("BEGIN PROGRAM").trim()
 }
 
 fun ExecuteFileTester.swiftTranslated(file: File): String {
@@ -67,18 +56,19 @@ fun ExecuteFileTester.swiftTranslated(file: File): String {
 }
 
 fun ExecuteFileTester.compileToSwift(file: File): File {
+    makeReady()
     val outFolder = swiftTestDir.resolve("Sources/testOut")
-    this.kotlinCompile(
-        sourceFile = file,
-        argumentsModification = {
-            this.pluginClasspaths = arrayOf("build/libs/kotlin-compiler-plugin-swift-0.2.0.jar")
-            this.pluginOptions =
-                arrayOf(
-                    "plugin:${KotlinSwiftCLP.PLUGIN_ID}:${KotlinTranspileCLP.KEY_EQUIVALENTS_NAME}=${swiftTestDir.resolve("Sources/KhrysalisRuntime")}",
-                    "plugin:${KotlinSwiftCLP.PLUGIN_ID}:${KotlinTranspileCLP.KEY_OUTPUT_DIRECTORY_NAME}=${outFolder}",
-                    "plugin:${KotlinSwiftCLP.PLUGIN_ID}:${KotlinTranspileCLP.KEY_PROJECT_NAME_NAME}=Yeet"
-                )
-        }
-    )
+    KotlinCompilation().apply {
+        inheritClassPath = true
+        sources = listOf(SourceFile.fromPath(file)) + Libraries.testingStubs.map { SourceFile.fromPath(it) }
+        commandLineProcessors = listOf(KotlinSwiftCLP())
+        compilerPlugins = listOf(KotlinSwiftCR())
+        pluginOptions = listOf(
+            PluginOption(KotlinSwiftCLP.PLUGIN_ID, KotlinTranspileCLP.KEY_EQUIVALENTS_NAME, swiftTestDir.resolve("Sources/KhrysalisRuntime").toString()),
+            PluginOption(KotlinSwiftCLP.PLUGIN_ID, KotlinTranspileCLP.KEY_OUTPUT_DIRECTORY_NAME, outFolder.toString()),
+            PluginOption(KotlinSwiftCLP.PLUGIN_ID, KotlinTranspileCLP.KEY_PROJECT_NAME_NAME, "Yeet"),
+            PluginOption(KotlinSwiftCLP.PLUGIN_ID, KotlinTranspileCLP.KEY_LIBRARY_MODE_NAME, "false"),
+        )
+    }.compile()
     return outFolder.resolve(file.nameWithoutExtension + ".swift")
 }
